src/test/java/ru/job4j/search/PhoneDictionaryTest.java
package ru.job4j.search;

import org.junit.Test;
import java.util.ArrayList;
import static org.hamcrest.core.Is.is;
import static org.junit.Assert.assertThat;

import static org.junit.Assert.*;

public class PhoneDictionaryTest {

    @Test
    public void whenFindByName() {
        var phones = new PhoneDictionary();
        phones.add(
                new Person("Petr", "Arsentev", "534872", "Bryansk")
        );
        var persons = phones.find("Petr");
        assertThat(persons.get(0).getSurname(), is("Arsentev"));
    }

    @Test
    public void whenFindByNameThenNothing() {
        var phones = new PhoneDictionary();
        phones.add(
                new Person("Oleg", "Arsentev", "534892", "Bryansk")
        );
        var persons = phones.find("line");
        assertTrue(persons.isEmpty());
    }
}